Use do..end for multiline blocks.

+++ app/models/hardware_server.rb
@@ -88,58 +88,50 @@ class HardwareServer < ActiveRecord::Base
 
     os_templates_list = os_templates_on_server.collect { |item| item.split[1] }
 
-    os_templates.each { |template|
-      if !os_templates_list.include?(template.name + '.tar.gz')
-        template.destroy
-      end
-    }
+    os_templates.each do |template|
+      template.destroy unless os_templates_list.include?(template.name + '.tar.gz')
+    end
 
-    os_templates_on_server.each { |template_record|
+    os_templates_on_server.each do |template_record|
       size, template_name = template_record.split
       template_name.sub!(/\.tar\.gz/, '')
 
       os_template = OsTemplate.find_or_create_by_name_and_hardware_server_id(template_name, self.id)
       os_template.size = size.to_i
       os_template.save
-    }
+    end
   end
 
   def sync_server_templates
     path = '/etc/vz/conf';
     server_templates_on_server = rpc_client.exec('ls', "#{path}/ve-*.conf-sample")['output'].split
 
-    server_templates.each { |template|
-      if !server_templates_on_server.include?("#{path}/ve-" + template.name + '.conf-sample')
-        template.destroy
-      end
-    }
+    server_templates.each do |template|
+      template.destroy unless server_templates_on_server.include?("#{path}/ve-" + template.name + '.conf-sample')
+    end
 
-    server_templates_on_server.each { |template_name|
+    server_templates_on_server.each do |template_name|
       template_name.sub!(/\/etc\/vz\/conf\/ve\-(.*)\.conf\-sample/, '\1')
       if !ServerTemplate.find_by_name_and_hardware_server_id(template_name, self.id)
         server_template = ServerTemplate.new(:name => template_name)
         server_template.hardware_server = self
         server_template.save
       end
-    }
+    end
   end
 
   def sync_virtual_servers
     ves_on_server = rpc_client.exec('vzlist', '-a -H -o veid,hostname,ip,status')['output'].split("\n")
     # skip error lines
-    ves_on_server = ves_on_server.find_all { |item| item =~ /^\s+\d+/ }
+    ves_on_server = ves_on_server.find_all{ |item| item =~ /^\s+\d+/ }
 
-    ves_ids_on_server = ves_on_server.map { |vzlist_entry|
-      vzlist_entry = vzlist_entry.split.first
-    }
+    ves_ids_on_server = ves_on_server.map{ |vzlist_entry| vzlist_entry = vzlist_entry.split.first }
 
-    virtual_servers.each { |virtual_server|
-      if !ves_ids_on_server.include?(virtual_server.identity.to_s)
-        virtual_server.destroy
-      end
-    }
+    virtual_servers.each do |virtual_server|
+      virtual_server.destroy unless ves_ids_on_server.include?(virtual_server.identity.to_s)
+    end
 
-    ves_on_server.each { |vzlist_entry|
+    ves_on_server.each do |vzlist_entry|
       ve_id, host_name, ip_address, ve_state = vzlist_entry.split
 
       virtual_server = virtual_servers.find_by_identity(ve_id)
@@ -171,7 +163,7 @@ class HardwareServer < ActiveRecord::Base
       end
 
       virtual_server.save(false)
-    }
+    end
   end
 
   def sync_backups
@@ -180,7 +172,7 @@ class HardwareServer < ActiveRecord::Base
     # remove totals line
     backups_list.shift
 
-    backups_list.each { |backup_record|
+    backups_list.each do |backup_record|
       size, filename = backup_record.split
       next unless match = filename.match(/^ve-dump\.(\d+)\.\d+.tar$/)
 
@@ -197,7 +189,7 @@ class HardwareServer < ActiveRecord::Base
 
       backup = Backup.new(:name => filename, :size => size.to_i, :virtual_server_id => virtual_server.id)
       backup.save
-    }
+    end
   end
 
   def sync_config
